+++ test/bevis/conformance_test.clj
@@ -3,8 +3,8 @@
             [bevis.test-support :as support]
             [clojure.test :refer [deftest]]))
 
-(deftest in-memory-adapter-obeys-challenge-contract
-  (conformance/assert-challenge-store (support/memory-adapter)))
+(deftest in-memory-store-obeys-challenge-contract
+  (conformance/assert-challenge-store (support/memory-store)))
 
-(deftest in-memory-adapter-obeys-session-contract
-  (conformance/assert-session-store (support/memory-adapter)))
+(deftest in-memory-store-obeys-session-contract
+  (conformance/assert-session-store (support/memory-store)))